+++ src/common/infra/file_list/postgres.rs
@@ -204,8 +204,10 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
                 }
             }
             // delete files by ids
-            let sql = format!("DELETE FROM file_list WHERE id IN({});", ids.join(","));
-            _ = pool.execute(sql.as_str()).await?;
+            if !ids.is_empty() {
+                let sql = format!("DELETE FROM file_list WHERE id IN({});", ids.join(","));
+                _ = pool.execute(sql.as_str()).await?;
+            }
         }
         Ok(())
     }
@@ -283,11 +285,13 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
                 }
             }
             // delete files by ids
-            let sql = format!(
-                "DELETE FROM file_list_deleted WHERE id IN({});",
-                ids.join(",")
-            );
-            _ = pool.execute(sql.as_str()).await?;
+            if !ids.is_empty() {
+                let sql = format!(
+                    "DELETE FROM file_list_deleted WHERE id IN({});",
+                    ids.join(",")
+                );
+                _ = pool.execute(sql.as_str()).await?;
+            }
         }
         Ok(())
     }
